[finished #162404814] fix the comment issue, readme and the merge the models
