+++ app/api/v1/views_users.py
@@ -1,6 +1,6 @@
 from flask_restful import Resource, request
 from flask import make_response, jsonify
-from .models_users import UserModel
+from .models import UserModel
 
 
 class CreateUser(Resource, UserModel):
