+++ app/tests/test_v1/__init__.py
@@ -1,11 +1,9 @@
-import json
+
 import unittest
-import pytest
+
 
 from app import create_app
-from app.api.v1 import api
 
-from app.api.v1.models_records import RedFlagRecordsModel as db
 
 from run import app
